Documentation controller jpa and exception

+++ src/main/java/pl/sda/jira/documentation/domain/DocumentationService.java
@@ -1,5 +1,6 @@
 package pl.sda.jira.documentation.domain;
 
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import pl.sda.jira.documentation.domain.exception.DocumentDoestExist;
 import pl.sda.jira.documentation.dto.DocumentationDto;
@@ -8,6 +9,7 @@ import java.util.UUID;
 
 @Service
 public class DocumentationService {
+
     private final DocumentationRepository documentationRepository;
 
     public DocumentationService(DocumentationRepository documentationRepository) {
@@ -23,15 +25,19 @@ public class DocumentationService {
     }
 
     public Long add(DocumentationDto documentationDto) {
-        Long documentationId = UUID.randomUUID().getMostSignificantBits();
-        Documentation documentation = new Documentation(documentationId, documentationDto.getTitle());
-        documentationRepository.add(documentation);
-        return documentationId;
+
+        Documentation documentation = new Documentation(documentationDto.getTitle());
+        Documentation documentation1 = documentationRepository.add(documentation);
+
+        return documentation1.getId();
     }
 
     public Long delete(Long documentationId) {
-        documentationRepository.delete(documentationId);
-        return documentationId;
+        if (exists(documentationId)) {
+            documentationRepository.delete(documentationId);
+            return documentationId;
+        }
+        throw new DocumentDoestExist(documentationId);
 
     }
